Move Minimal Join Flow into its own file
