+++ packages/join-flow/src/app.tsx
@@ -19,14 +19,15 @@ import {
 import { useOnce } from "./hooks/util";
 import { PaymentDetailsPage } from "./pages/payment-details.page";
 import { Stager } from "./components/stager";
-import { currencyCodeToSymbol, FormSchema, getTestDataIfEnabled } from "./schema";
+import { FormSchema, getTestDataIfEnabled } from "./schema";
 import { ConfirmationPage } from "./pages/confirm.page";
 import { get as getEnv, getPaymentMethods } from "./env";
 import { usePostResource } from "./services/rest-resource.service";
 import gocardless from "./images/gocardless.svg";
 import chargebee from "./images/chargebee.png";
 
-import { Elements, PaymentElement, useStripe, useElements } from '@stripe/react-stripe-js';
+import { Elements } from '@stripe/react-stripe-js';
+import MinimalJoinForm from "./components/minimal-join-flow";
 import { loadStripe } from '@stripe/stripe-js';
 
 interface Stage {
@@ -167,22 +168,9 @@ const App = () => {
       </a>
   })
 
-  // mode is going to have to be payment
-
   const options = {
-    mode: 'subscription',
-    amount: 1000,
-    currency: 'gbp',
     paymentMethodCreation: 'manual',
-    // Fully customizable with appearance API.
-    appearance: {/*...*/ },
   };
-
-  /*
-  Start £4 a month, end at £40 a month
-
-  Donations start at £10 and end at £100
-  */
 
   // @ts-ignore
   const minimalJoinForm = <Elements stripe={stripePromise} options={options}>
@@ -303,118 +291,5 @@ const Fail: FC<{ router: StateRouter }> = ({ router }) => {
 
   return null;
 };
-
-const MinimalJoinForm = () => {
-  const stripe = useStripe();
-  const elements = useElements();
-
-  const [errorMessage, setErrorMessage] = useState();
-  const [otherMessage, setOtherMessage] = useState();
-  const [loading, setLoading] = useState(false);
-
-  const [selectedPlan, setSelectedPlan] = useState(null);
-  const [plans, setPlans] = useState([]);
-  const [email, setEmail] = useState('');
-
-  const handleError = (error: { message: string }) => {
-    setLoading(false);
-    setErrorMessage(error.message);
-  }
-
-  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    if (!stripe) {
-      return;
-    }
-
-    setLoading(true);
-
-    const { error: submitError } = await elements!.submit();
-
-    if (submitError) {
-      handleError(submitError);
-      return;
-    }
-
-    const { error, confirmationToken } = await stripe!.createConfirmationToken({
-      elements
-    });
-
-    // This point is only reached if there's an immediate error when
-    // creating the ConfirmationToken. Show the error to your customer (for example, payment details incomplete)
-    if (error) {
-      handleError(error);
-      return;
-    }
-
-    // Pass the confirmation token to the server to create a subscription
-    const APIEndpoint = getEnv('WP_REST_API') + 'join/v1/stripe/create-confirm-subscription';
-
-    const res = await fetch(APIEndpoint, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        confirmationTokenId: confirmationToken.id,
-        email,
-        membership: selectedPlan?.label
-      }),
-    });
-
-    const data = await res.json();
-
-    if (data.status !== 'succeeded') {
-      setErrorMessage('Connection to payment provider failed, please try again');
-    }
-
-    setOtherMessage('Payment successful, welcome');
-  };
-
-  const handleRangeChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = parseInt(e.target.value);;
-
-    setSelectedPlan(plans[value]);
-  };
-
-  useEffect(() => {
-    const plans = getEnv("MEMBERSHIP_PLANS");
-
-    const permissableMembershipPlans = plans.filter((plan) => !plan.allowCustomAmount);
-    const sortedPlans = permissableMembershipPlans.sort((a, b) => a.amount - b.amount);
-    const medianIndex = Math.floor(sortedPlans.length / 2);
-    const medianPlan = sortedPlans[medianIndex];
-
-    setSelectedPlan(medianPlan);
-    setPlans(sortedPlans);
-  }, []);
-
-  const handleEmailChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setEmail(e.target.value);
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <div>
-        {selectedPlan && `${currencyCodeToSymbol(selectedPlan.currency)}${selectedPlan.amount} ${selectedPlan.frequency}`}
-      </div>
-      <input
-        type="range"
-        min={0}
-        max={plans.length - 1}
-        step="1"
-        onChange={handleRangeChange}
-        value={plans.findIndex(plan => plan === selectedPlan)}
-      />
-      <div>
-        <label htmlFor="email">Email</label>
-        <input type="email" name="email" value={email} onChange={handleEmailChange}></input>
-        <PaymentElement />
-        <button type="submit" disabled={!stripe || loading}>Pay Now</button>
-        {errorMessage && <div>{errorMessage}</div>}
-        {otherMessage && <div>{otherMessage}</div>}
-      </div>
-    </form>
-  );
-}
 
 export default App;
